agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Local voice assistant with Pipecat, n8n workflows, and MCP integrations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = [
    "pipecat",
    "voice",
    "speech-to-text",
    "text-to-speech",
    "mcp",
    "n8n",
    "ollama",
    "local",
    "self-hosted",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Pipecat core + services
    "pipecat-ai[silero,riva,ollama]>=0.0.80",

    # MCP client
    "mcp>=1.0.0",

    # Core utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",

    # Webhook server
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",

    # Web search
    "duckduckgo-search>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a1",
]

[project.scripts]
voice-agent = "voice_agent.bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["voice_agent"]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["voice_agent"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
